[refactor] Order 도메인 역할 분리
