+++ src/main/java/com/bbangle/bbangle/order/seller/controller/SellerOrderApi.java
@@ -1,18 +1,18 @@
-package com.bbangle.bbangle.order.controller.swagger;
+package com.bbangle.bbangle.order.seller.controller;
 
 import com.bbangle.bbangle.common.dto.ListResult;
 import com.bbangle.bbangle.common.dto.SingleResult;
 import com.bbangle.bbangle.common.page.BbanglePageResponse;
 import com.bbangle.bbangle.exception.ErrorResponse;
 import com.bbangle.bbangle.exception.GlobalControllerAdvice;
-import com.bbangle.bbangle.order.controller.dto.request.CompletedOrderFilter;
-import com.bbangle.bbangle.order.controller.dto.request.OrderRequest;
-import com.bbangle.bbangle.order.controller.dto.request.SellerOrderRequest;
-import com.bbangle.bbangle.order.controller.dto.response.CompletedOrderResponse.OrderSummary;
-import com.bbangle.bbangle.order.controller.dto.response.OrderDetailResponse.OrderDetail;
-import com.bbangle.bbangle.order.controller.dto.response.OrderResponse.OrderItemDetailResponse;
-import com.bbangle.bbangle.order.controller.dto.response.OrderResponse.OrderSearchResponse;
-import com.bbangle.bbangle.order.controller.dto.response.SellerOrderResponse.OrderConfirmResponse;
+import com.bbangle.bbangle.order.seller.controller.dto.request.CompletedOrderFilter;
+import com.bbangle.bbangle.order.seller.controller.dto.request.OrderRequest;
+import com.bbangle.bbangle.order.seller.controller.dto.request.SellerOrderRequest;
+import com.bbangle.bbangle.order.seller.controller.dto.response.CompletedOrderResponse.OrderSummary;
+import com.bbangle.bbangle.order.seller.controller.dto.response.OrderDetailResponse.OrderDetail;
+import com.bbangle.bbangle.order.seller.controller.dto.response.OrderResponse.OrderItemDetailResponse;
+import com.bbangle.bbangle.order.seller.controller.dto.response.OrderResponse.OrderSearchResponse;
+import com.bbangle.bbangle.order.seller.controller.dto.response.SellerOrderResponse.OrderConfirmResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
